Calculate total NEWS score from individual components

Calculate total NEWS score from individual components

Usage

news_total(hr, rf, temp, sbp, avpu, gcs, o2supp, spo2, return_df = FALSE)

Arguments

hr

Numeric value representing Heart Rate (bpm)

rf

Numeric value representing Respiratory Frequency (breaths per minute)

temp

Numeric value representing Temperature (°C)

sbp

Numeric value representing Systolic Blood Pressure (mmHg)

avpu

Character value representing AVPU scale ("A", "V", "P", "U", "D")

gcs

Numeric value representing Glasgow Coma Scale (3-15)

o2supp

Logical value indicating if oxygen supplementation is provided (TRUE/FALSE)

spo2

Numeric value representing Oxygen Saturation (%)

return_df

Logical value indicating if a detailed dataframe should be returned (TRUE/FALSE), default is FALSE

Value

Total NEWS score as an integer, or a dataframe with individual component scores and total score if return_df is TRUE

See also